refactor: 과하게 구체적이던 패키지명 변경
