+++ src/main/java/roomescape/auth/AuthService.java
@@ -4,8 +4,8 @@ import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.security.Keys;
 import org.springframework.dao.IncorrectResultSizeDataAccessException;
 import org.springframework.stereotype.Service;
-import roomescape.auth.token.jwt.JwtProperties;
-import roomescape.auth.token.jwt.JwtProvider;
+import roomescape.auth.client.jwt.JwtProperties;
+import roomescape.auth.client.jwt.JwtProvider;
 import roomescape.global.exception.RoomescapeUnauthorizedException;
 import roomescape.member.Member;
 import roomescape.member.MemberDao;
